+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Preserve filters across task types (#265)
 -   Explicit messages for why models cannot be downloaded (#263)
 -   New help & feedback modal (#273)
+-   Add support for Microsoft Clarity (optional) with a cookie banner (#241)
 
 ### Fixed